% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotyping_target_lookup.R
\name{create_target_BEDfile}
\alias{create_target_BEDfile}
\title{Create a New Entry on the Master Bedfile}
\usage{
create_target_BEDfile(
  query,
  crispr_target_name,
  genome = c("Drerio", "GFP"),
  master_bed = "data-raw/master.bed"
)
}
\arguments{
\item{query}{the guide RNA sequence, using DNA nucleotides}

\item{crispr_target_name}{The name for the crispr site.}
}
\value{
Nothing.  Adds an entry to the master bedfile.
}
\description{
Use this function to find a crispr target sequence and write to the master bedfile.  Query should be a 23-letter string corresponding to the gRNA used to make the cripsr, entered 5' to 3'.  The target recognition site should first, followed by the PAM.  By convention, all crispr target names will be converted to upper case.  If either the target range or name are duplicated, an error will result.
}
\seealso{
\code{\link[rtracklayer]{character(0)}}
\code{\link[Biostrings]{DNAString-class}},\code{\link[Biostrings]{matchPattern}}
\code{\link[BSgenome.Drerio.blaserlabgenotyping.dr11]{BSgenome.Drerio.blaserlabgenotyping.dr11}}
}
