% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotyping_main.R
\name{genotyping_main}
\alias{genotyping_main}
\title{Genotype a Crispr Locus}
\usage{
genotyping_main(
  crispr_target,
  genome = c("Drerio", "GFP"),
  network_directory,
  multiplex = TRUE,
  read_threshold = 25,
  bed = "data-raw/master.bed",
  split_snv = TRUE,
  txdb = system.file("extdata/GRCz11.97_txdb.sqlite", package =
    "BSgenome.Drerio.blaserlabgenotyping.dr11"),
  cutadapt_path = "/workspace/python/anaconda3/envs/cutadaptenv/bin/cutadapt"
)
}
\arguments{
\item{crispr_target}{The name of the crispr target locus.  Must be present on the master BEDfile}

\item{genome}{The genome to use.  Can be either Drerio or GFP.  Default is Drerio.}

\item{network_directory}{Network directory containing fastqs.  The fastqs can be nested in a subdirectory; the output will be placed in the directory named.}

\item{multiplex}{Whether or not to demultiplex, Default: TRUE}

\item{read_threshold}{Remove alleles with read counts below this threshold, Default: 25}

\item{bed}{master bedfile to use, Default: 'data-raw/master.bed'}

\item{split_snv}{If TRUE, will count SNVs as variant alleles.}

\item{txdb}{transcriptome database to use.  If using a nonstandard transgene, should be set to NULL.}

\item{cutadapt_path}{The path to your cutadapt binary.  Find this by running $ which cutadapt in your terminal.}
}
\value{
Nothing; writes out to the network directory.
}
\description{
Use this function to genotype a target crispr locus.
}
\seealso{
\code{\link[rtracklayer]{character(0)}}
\code{\link[parallel]{detectCores}}
}
