% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotyping_main.R
\name{genotyping_main}
\alias{genotyping_main}
\title{Genotype a Crispr Locus}
\usage{
genotyping_main(
  crispr_target,
  network_directory,
  multiplex = TRUE,
  read_threshold = 25,
  bed = "data-raw/master.bed"
)
}
\arguments{
\item{crispr_target}{The name of the crispr target locus.  Must be present on the master BEDfile}

\item{network_directory}{Network directory containing fastqs.  The fastqs can be nested in a subdirectory; the output will be placed in the directory named.}

\item{multiplex}{Whether or not to demultiplex, Default: TRUE}

\item{read_threshold}{Remove alleles with read counts below this threshold, Default: 25}

\item{bed}{master bedfile to use, Default: 'data-raw/master.bed'}
}
\value{
Nothing; writes out to the network directory.
}
\description{
Use this function to genotype a target crispr locus.
}
\seealso{
\code{\link[rtracklayer]{character(0)}}
\code{\link[parallel]{detectCores}}
}
